+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Changelog
 - Removed `dukt\twitter\helpers\TwitterHelper::formatTime()` unused method.
 - Removed `dukt\twitter\web\twig\variables\TwitterVariable::getUserImageUrl()` deprecated method.
 - Removed `dukt\twitter\web\twig\Extension::embedTweet()` deprecated method.
+- Removed `dukt\twitter\services\Twitter::embedTweet()` deprecated method.
 
 ### Fixed
 - Fixed a bug where the Tweet field would throw an error when trying to show a deleted tweet. ([#7](https://github.com/dukt/twitter/issues/7))